requirements.txt
faiss-cpu
numpy
requests
PySide6
